Split up a bunch of files
