+++ src/ts/core/common/selection.ts
@@ -1,0 +1,1 @@
+export const unselectText = () => window.getSelection()?.removeAllRanges()
