+++ src/ts/core/features/index.ts
@@ -6,7 +6,7 @@ import {config as incDec} from './inc-dec-value'
 import {config as srs} from '../srs/srs'
 import {config as blockManipulation} from './block-manipulation'
 import {config as vimMode} from './vim-mode'
-import {config as spatialGraphMode} from './spatial-graph-mode'
+import {config as spatialMode} from './spatial-mode'
 import {config as estimate} from './estimates'
 import {config as navigation} from './navigation'
 import {config as livePreview} from './livePreview'
@@ -23,7 +23,7 @@ export const Features = {
         srs,
         blockManipulation,
         vimMode,
-        spatialGraphMode,
+        spatialMode,
         estimate,
         navigation,
         dateTitle,
